+++ scripts/v120_plot.py
@@ -8,7 +8,7 @@ import matplotlib.cm as cm
 curve = Bezier([[0, 0, 0], [3, 6, 0], [5, -3, 0], [10, 0, 0]])
 
 viewer = App(viewmode="shaded", enable_sidebar=True, width=1600, height=900)
-pointobj = viewer.add(Point(* curve.point(0)), size=20, color=(1, 0, 0))
+pointobj = viewer.add(Point(* curve.point(0)), pointsize=20, pointcolor=(1, 0, 0))
 curveobj = viewer.add(Polyline(curve.locus()), linewidth=2)
 
 figure = viewer.plot("My Plot", location="bottom", min_height=300)