+++ scripts/v120_ui_elements2.py
@@ -5,15 +5,14 @@ from compas_view2.app import App
 curve = Bezier([[0, 0, 0], [3, 6, 0], [5, -3, 0], [10, 0, 0]])
 
 viewer = App(viewmode="shaded", enable_sidebar=True, width=1600, height=900)
-viewer.view.camera.tx = -5.0
-viewer.view.camera.rz = 0
-viewer.view.camera.rx = -20
+viewer.view.camera.target = [5, 0, 0]
+viewer.view.camera.distance = 20
 
 viewer.add(Polyline(curve.locus()), linewidth=2)
 
 points = []
 for i in range(10):
-    p = viewer.add(Point(* curve.point(0)), size=20, color=(1, 0, 0))
+    p = viewer.add(Point(* curve.point(0)), pointsize=20, pointcolor=(1, 0, 0))
     points.append(p)