+++ src/compas_view2/app/app.py
@@ -252,22 +252,73 @@ class App:
         y = 0.5 * (rect.height() - height)
         self.window.setGeometry(x, y, width, height)
 
-    def add(self, data: Data, **kwargs) -> Object:
+    def add(self, data: Data,
+            name: str = None,
+            is_selected: bool = False,
+            is_visible: bool = True,
+            show_points: bool = False,
+            show_lines: bool = True,
+            show_faces: bool = True,
+            pointcolor: Union[Color, Dict[Union[str, int], Color]] = None,
+            linecolor: Union[Color, Dict[Union[str, int], Color]] = None,
+            facecolor: Union[Color, Dict[Union[str, int], Color]] = None,
+            linewidth: int = 1,
+            pointsize: int = 10,
+            opacity: int = 1.0,
+            **kwargs) -> Object:
         """Add a COMPAS object.
 
         Parameters
         ----------
-        data: :class:`compas.geometry.Primitive` | :class:`compas.geometry.Shape` | :class:`compas.datastructures.Datastructure`
+        data: :class:`compas.data.Data`
             A COMPAS data object.
+        is_selected : bool, optional
+            Whether the object is selected.
+            Default to False.
+        is_visible : bool, optional
+            Whether to show object.
+            Default to True.
+        show_points : bool, optional
+            Whether to show points/vertices of the object.
+            Default to False.
+        show_lines : bool, optional
+            Whether to show lines/edges of the object.
+            Default to True.
+        show_faces : bool, optional
+            Whether to show faces of the object.
+            Default to True.
+        pointcolor : Union[Color, Dict[Union[str, int], Color]], optional
+            The color or the dict of colors of the points.
+            Default to `compas_view2.objects.Object.default_color_points`.
+        linecolor : Union[Color, Dict[Union[str, int], Color]], optional
+            The color or the dict of colors of the lines.
+            Default to `compas_view2.objects.Object.default_color_lines`.
+        facecolor : Union[Color, Dict[Union[str, int], Color]], optional
+            The color or the dict of colors of the faces.
+            Default to `compas_view2.objects.Object.default_color_faces`.
+        linewidth : int, optional
+            The line width to be drawn on screen
+            Default to 1.
+        pointsize : int, optional
+            The point size to be drawn on screen
+            Default to 10.
+        opacity : float, optional
+            The opacity of the object.
+            Default to 1.0.
         **kwargs : dict, optional
-            Additional visualization options.
+            Additional visualization options for specific objects.
 
         Returns
         -------
         :class:`compas_view2.objects.Object`
+            The added object.
 
         """
-        obj = Object.build(data, app=self, **kwargs)
+        obj = Object.build(data, name=name, is_selected=is_selected, is_visible=is_visible,
+                           show_points=show_points, show_lines=show_lines, show_faces=show_faces,
+                           pointcolor=pointcolor, linecolor=linecolor, facecolor=facecolor,
+                           linewidth=linewidth, pointsize=pointsize, opacity=opacity, app=self, **kwargs)
+
         self.view.objects[obj] = obj
         self.selector.add(obj)
         if self.view.isValid():