+++ CHANGELOG.md
@@ -21,12 +21,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-* `treeform` is now more customizable.
+* Moved `pointcolor`, `linecolor`, `facecolor` to `Object` from `BufferObject`.
+* Exposing `pointcolor`, `linecolor`, `facecolor` at `App.add` for pylance linting.
+* `pointcolor`, `linecolor`, `facecolor` now accept both single value or dictionary of values.
 
 ### Removed
 
+* `pointcolors`, `linecolors`, `facecolors` are removed, use `pointcolor`, `linecolor`, `facecolor` instead.
+* `treeform` is now more customizable.
 * Removed placeholder menu items.
-
 
 ## [0.6.0] 2022-07-21